move nft types into separate file
