+++ src/util/inputSanitization.ts
@@ -1,5 +1,5 @@
 import { InputNftTokenType } from '../internal/nft-api';
-import { NftTokenType } from '../types/types';
+import { NftTokenType } from '../types/nft-types';
 
 export function sanitizeTokenType(tokenType?: NftTokenType): InputNftTokenType {
   if (tokenType === NftTokenType.ERC1155 || tokenType === NftTokenType.ERC721) {
